Fix regex search qualifiers (#10)
* Fix regex for in: search qualifier

After exchanging emails with GitHub support, it appears that I was mistaken,
and the `in:filename` search qualifier doesn't exist. `in:path`, `in:file`,
and `in:file,path` do though.

* Detect extension: search qualifier anywhere in search query

The `extension:` search qualifier may not necessarily be at the beginning of
the search query. We don't need to add it if we find it anywhere in the
search query, same as for e.g., `filename:`.